requirements.txt
fastapi
uvicorn
selenium>=4.6.0
beautifulsoup4
pandas
requests
lxml
google-api-python-client
pyairtable
python-dotenv
pydantic-settings
